requirements.txt
Flask
mysql-connector-python
qrcode
Pillow
requests
